Refactor plugin structure and update package names for consistency
