+++ skyhigh-16kb-doctor/plugin/src/functionalTest/kotlin/com/sparrow/plugin/PluginBuildPluginFunctionalTest.kt
@@ -1,7 +1,7 @@
 /*
  * This source file was generated by the Gradle 'init' task
  */
-package org.example
+package org.sparrow.plugin
 
 import java.io.File
 import kotlin.test.assertTrue
