+++ skyhigh-16kb-doctor/plugin/src/main/kotlin/com/sparrow/plugin/PluginBuildPlugin.kt
@@ -1,7 +1,7 @@
 /*
  * This source file was generated by the Gradle 'init' task
  */
-package org.example
+package com.sparrow.plugin
 
 import org.gradle.api.Project
 import org.gradle.api.Plugin
