+++ skyhigh-16kb-doctor/plugin/src/test/kotlin/com/sparrow/plugin/PluginBuildPluginTest.kt
@@ -1,7 +1,7 @@
 /*
  * This source file was generated by the Gradle 'init' task
  */
-package org.example
+package org.sparrow.plugin
 
 import org.gradle.testfixtures.ProjectBuilder
 import kotlin.test.Test
